Update packages in lambda
Issue #6 suggests updating clamav, which this does.

It is a bit more involved though:
The base image installs the yum plugin "priorities", which is used to pin
packages (including clamav) to the base image information.
To work around that, remove the plugin first.

We'll then update all packages, and then cleanup after ourselves.